requirements.txt
ultralytics
streamlit
google-generativeai
torch
numpy
Pillow
opencv-python
gtts
